+++ docs/source/conf.py
@@ -53,13 +53,13 @@ autodoc_mock_imports = [
 
 # General information about the project.
 project = "OSMnx"
-copyright = "2016–2021, Geoff Boeing"
+copyright = "2016–2022, Geoff Boeing"
 author = "Geoff Boeing"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
-version = release = "1.1.2"
+version = release = "1.2.0"
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #