+++ osmnx/io.py
@@ -7,7 +7,6 @@ import networkx as nx
 import pandas as pd
 from shapely import wkt
 
-from . import osm_xml
 from . import settings
 from . import utils
 from . import utils_graph
@@ -167,25 +166,33 @@ def save_graphml(G, filepath=None, gephi=False, encoding="utf-8"):
     utils.log(f'Saved graph as GraphML file at "{filepath}"')
 
 
-def load_graphml(filepath, node_dtypes=None, edge_dtypes=None, graph_dtypes=None):
+def load_graphml(
+    filepath=None, graphml_str=None, node_dtypes=None, edge_dtypes=None, graph_dtypes=None
+):
     """
-    Load an OSMnx-saved GraphML file from disk.
+    Load an OSMnx-saved GraphML file from disk or GraphML string.
 
-    This converts node, edge, and graph-level attributes (serialized as
-    strings) to their appropriate data types. These can be customized as
+    This function converts node, edge, and graph-level attributes (serialized
+    as strings) to their appropriate data types. These can be customized as
     needed by passing in dtypes arguments providing types or custom converter
     functions. For example, if you want to convert some attribute's values to
     `bool`, consider using the built-in `ox.io._convert_bool_string` function
     to properly handle "True"/"False" string literals as True/False booleans:
-    `ox.load_graphml(fp, node_dtypes={my_attr: ox.io._convert_bool_string})`
+    `ox.load_graphml(fp, node_dtypes={my_attr: ox.io._convert_bool_string})`.
 
     If you manually configured the `all_oneway=True` setting, you may need to
     manually specify here that edge `oneway` attributes should be type `str`.
+
+    Note that you must pass one and only one of `filepath` or `graphml_str`.
+    If passing `graphml_str`, you may need to decode the bytes read from your
+    file before converting to string to pass to this function.
 
     Parameters
     ----------
     filepath : string or pathlib.Path
         path to the GraphML file
+    graphml_str : string
+        a valid and decoded string representation of a GraphML file's contents
     node_dtypes : dict
         dict of node attribute names:types to convert values' data types. the
         type can be a python type or a custom string converter function.
@@ -201,7 +208,10 @@ def load_graphml(filepath, node_dtypes=None, edge_dtypes=None, graph_dtypes=None
     -------
     G : networkx.MultiDiGraph
     """
-    filepath = Path(filepath)
+    if (filepath is None and graphml_str is None) or (
+        filepath is not None and graphml_str is not None
+    ):  # pragma: no cover
+        raise ValueError("You must pass one and only one of `filepath` or `graphml_str`.")
 
     # specify default graph/node/edge attribute values' data types
     default_graph_dtypes = {"simplified": _convert_bool_string}
@@ -222,6 +232,7 @@ def load_graphml(filepath, node_dtypes=None, edge_dtypes=None, graph_dtypes=None
         "length": float,
         "oneway": _convert_bool_string,
         "osmid": int,
+        "reversed": _convert_bool_string,
         "speed_kph": float,
         "travel_time": float,
     }
@@ -234,8 +245,18 @@ def load_graphml(filepath, node_dtypes=None, edge_dtypes=None, graph_dtypes=None
     if edge_dtypes is not None:
         default_edge_dtypes.update(edge_dtypes)
 
-    # read the graphml file from disk
-    G = nx.read_graphml(filepath, node_type=default_node_dtypes["osmid"], force_multigraph=True)
+    if filepath is not None:
+        # read the graphml file from disk
+        source = filepath
+        G = nx.read_graphml(
+            Path(filepath), node_type=default_node_dtypes["osmid"], force_multigraph=True
+        )
+    else:
+        # parse the graphml string
+        source = "string"
+        G = nx.parse_graphml(
+            graphml_str, node_type=default_node_dtypes["osmid"], force_multigraph=True
+        )
 
     # convert graph/node/edge attribute data types
     utils.log("Converting node, edge, and graph-level attribute data types")
@@ -243,7 +264,7 @@ def load_graphml(filepath, node_dtypes=None, edge_dtypes=None, graph_dtypes=None
     G = _convert_node_attr_types(G, default_node_dtypes)
     G = _convert_edge_attr_types(G, default_edge_dtypes)
 
-    utils.log(f'Loaded graph with {len(G)} nodes and {len(G.edges)} edges from "{filepath}"')
+    utils.log(f'Loaded graph with {len(G)} nodes and {len(G.edges)} edges from "{source}"')
     return G
 
 
@@ -390,73 +411,3 @@ def _stringify_nonnumeric_cols(gdf):
             gdf[col] = gdf[col].fillna("").astype(str)
 
     return gdf
-
-
-def save_graph_xml(
-    data,
-    filepath=None,
-    node_tags=settings.osm_xml_node_tags,
-    node_attrs=settings.osm_xml_node_attrs,
-    edge_tags=settings.osm_xml_way_tags,
-    edge_attrs=settings.osm_xml_way_attrs,
-    oneway=False,
-    merge_edges=True,
-    edge_tag_aggs=None,
-):
-    """
-    Do not use: deprecated. Use osm_xml.save_graph_xml instead.
-
-    Parameters
-    ----------
-    data : networkx multi(di)graph OR a length 2 iterable of nodes/edges
-        geopandas GeoDataFrames
-    filepath : string or pathlib.Path
-        path to the .osm file including extension. if None, use default data
-        folder + graph.osm
-    node_tags : list
-        osm node tags to include in output OSM XML
-    node_attrs: list
-        osm node attributes to include in output OSM XML
-    edge_tags : list
-        osm way tags to include in output OSM XML
-    edge_attrs : list
-        osm way attributes to include in output OSM XML
-    oneway : bool
-        the default oneway value used to fill this tag where missing
-    merge_edges : bool
-        if True merges graph edges such that each OSM way has one entry
-        and one entry only in the OSM XML. Otherwise, every OSM way
-        will have a separate entry for each node pair it contains.
-    edge_tag_aggs : list of length-2 string tuples
-        useful only if merge_edges is True, this argument allows the user
-        to specify edge attributes to aggregate such that the merged
-        OSM way entry tags accurately represent the sum total of
-        their component edge attributes. For example, if the user
-        wants the OSM way to have a "length" attribute, the user must
-        specify `edge_tag_aggs=[('length', 'sum')]` in order to tell
-        this method to aggregate the lengths of the individual
-        component edges. Otherwise, the length attribute will simply
-        reflect the length of the first edge associated with the way.
-
-    Returns
-    -------
-    None
-    """
-    import warnings
-
-    msg = (
-        "The save_graph_xml function has been moved to the osm_xml module and "
-        "will be removed in a future release. Use the osm_xml module instead."
-    )
-    warnings.warn(msg)
-    osm_xml.save_graph_xml(
-        data,
-        filepath,
-        node_tags,
-        node_attrs,
-        edge_tags,
-        edge_attrs,
-        oneway,
-        merge_edges,
-        edge_tag_aggs,
-    )