+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change log
 
+## 1.2.1 (TBD)
+
+  - fix rate limit checking and pausing on newest versions of Overpass API
+  - allow add_edge_lengths function to be run on a subset of edges
+
+## 1.2.0 (2022-05-23)
+
+  - add ability to load GraphML string data to the load_graphml function
+  - add "reversed" edge attribute to support node-order-dependent edge attributes
+  - add new edge_color and edge_linewidth arguments to plot_footprints function
+  - fix nearest_edges function selecting arbitrary edge when bounding boxes overlap
+  - fix get_digraph function's parallel edge handling
+  - fix pandas and geopandas version compatibility
+  - fix log output appearing in Jupyter notebooks on Unix-like systems
+  - remove old functions and arguments previously deprecated in v1.1
+  - deprecate utils.config function in favor of using settings module directly
+
 ## 1.1.2 (2021-11-17)
   
   - fix geocoding when no geojson is returned